+++ libs/langchain_v1/tests/unit_tests/agents/middleware/core/test_wrap_model_call.py
@@ -1,18 +1,30 @@
-"""Unit tests for wrap_model_call middleware generator protocol."""
+"""Unit tests for wrap_model_call hook and @wrap_model_call decorator.
+
+This module tests the wrap_model_call functionality in three forms:
+1. As a middleware method (AgentMiddleware.wrap_model_call)
+2. As a decorator (@wrap_model_call)
+3. Async variant (AgentMiddleware.awrap_model_call)
+"""
+
+from collections.abc import Awaitable, Callable
 
 import pytest
+from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_core.language_models.fake_chat_models import GenericFakeChatModel
 from langchain_core.messages import AIMessage, HumanMessage
+from langchain_core.outputs import ChatGeneration, ChatResult
 
 from langchain.agents import create_agent
 from langchain.agents.middleware.types import (
     AgentMiddleware,
-    AgentState,
     ModelRequest,
+    wrap_model_call,
 )
 
+from ...model import FakeToolCallingModel
 
-class TestBasicOnModelCall:
+
+class TestBasicWrapModelCall:
     """Test basic wrap_model_call functionality."""
 
     def test_passthrough_middleware(self) -> None:
@@ -70,7 +82,7 @@ class TestBasicOnModelCall:
         assert counter.call_count == 1
 
 
-class TestRetryMiddleware:
+class TestRetryLogic:
     """Test retry logic with wrap_model_call."""
 
     def test_simple_retry_on_error(self) -> None:
@@ -91,12 +103,10 @@ class TestRetryMiddleware:
 
             def wrap_model_call(self, request, handler):
                 try:
-                    result = handler(request)
-                    return result
+                    return handler(request)
                 except Exception:
                     self.retry_count += 1
-                    result = handler(request)
-                    return result
+                    return handler(request)
 
         retry_middleware = RetryOnceMiddleware()
         model = FailOnceThenSucceed(messages=iter([AIMessage(content="Success")]))
@@ -125,8 +135,7 @@ class TestRetryMiddleware:
                 for attempt in range(self.max_retries):
                     self.attempts.append(attempt + 1)
                     try:
-                        result = handler(request)
-                        return result
+                        return handler(request)
                     except Exception as e:
                         last_exception = e
                         continue
@@ -142,6 +151,75 @@ class TestRetryMiddleware:
             agent.invoke({"messages": [HumanMessage("Test")]})
 
         assert retry_middleware.attempts == [1, 2, 3]
+
+    def test_no_retry_propagates_error(self) -> None:
+        """Test that error is propagated when middleware doesn't retry."""
+
+        class FailingModel(BaseChatModel):
+            """Model that always fails."""
+
+            def _generate(self, messages, **kwargs):
+                raise ValueError("Model error")
+
+            @property
+            def _llm_type(self):
+                return "failing"
+
+        class NoRetryMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                return handler(request)
+
+        agent = create_agent(model=FailingModel(), middleware=[NoRetryMiddleware()])
+
+        with pytest.raises(ValueError, match="Model error"):
+            agent.invoke({"messages": [HumanMessage("Test")]})
+
+    def test_max_attempts_limit(self) -> None:
+        """Test that middleware controls termination via retry limits."""
+
+        class AlwaysFailingModel(BaseChatModel):
+            """Model that always fails."""
+
+            def _generate(self, messages, **kwargs):
+                raise ValueError("Always fails")
+
+            @property
+            def _llm_type(self):
+                return "always_failing"
+
+        class LimitedRetryMiddleware(AgentMiddleware):
+            """Middleware that limits its own retries."""
+
+            def __init__(self, max_retries: int = 10):
+                super().__init__()
+                self.max_retries = max_retries
+                self.attempt_count = 0
+
+            def wrap_model_call(self, request, handler):
+                last_exception = None
+                for attempt in range(self.max_retries):
+                    self.attempt_count += 1
+                    try:
+                        return handler(request)
+                    except Exception as e:
+                        last_exception = e
+                        # Continue to retry
+
+                # All retries exhausted, re-raise the last error
+                if last_exception:
+                    raise last_exception
+
+        model = AlwaysFailingModel()
+        middleware = LimitedRetryMiddleware(max_retries=10)
+
+        agent = create_agent(model=model, middleware=[middleware])
+
+        # Should fail with the model's error after middleware stops retrying
+        with pytest.raises(ValueError, match="Always fails"):
+            agent.invoke({"messages": [HumanMessage("Test")]})
+
+        # Should have attempted exactly 10 times as configured
+        assert middleware.attempt_count == 10
 
 
 class TestResponseRewriting:
@@ -185,6 +263,28 @@ class TestResponseRewriting:
 
         assert result["messages"][1].content == "[BOT]: Response"
 
+    def test_multi_stage_transformation(self) -> None:
+        """Test middleware applying multiple transformations."""
+
+        class MultiTransformMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                result = handler(request)
+                # result is ModelResponse, extract AIMessage from it
+                ai_message = result.result[0]
+
+                # First transformation: uppercase
+                content = ai_message.content.upper()
+                # Second transformation: add prefix and suffix
+                content = f"[START] {content} [END]"
+                return AIMessage(content=content)
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="hello")]))
+        agent = create_agent(model=model, middleware=[MultiTransformMiddleware()])
+
+        result = agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert result["messages"][1].content == "[START] HELLO [END]"
+
 
 class TestErrorHandling:
     """Test error handling with wrap_model_call."""
@@ -200,9 +300,8 @@ class TestErrorHandling:
             def wrap_model_call(self, request, handler):
                 try:
                     return handler(request)
-                except Exception:
-                    fallback = AIMessage(content="Error handled gracefully")
-                    return fallback
+                except Exception as e:
+                    return AIMessage(content=f"Error occurred: {e}. Using fallback response.")
 
         model = AlwaysFailModel(messages=iter([]))
         agent = create_agent(model=model, middleware=[ErrorToSuccessMiddleware()])
@@ -210,7 +309,8 @@ class TestErrorHandling:
         # Should not raise, middleware converts error to response
         result = agent.invoke({"messages": [HumanMessage("Test")]})
 
-        assert "Error handled gracefully" in result["messages"][1].content
+        assert "Error occurred" in result["messages"][1].content
+        assert "fallback response" in result["messages"][1].content
 
     def test_selective_error_handling(self) -> None:
         """Test middleware that only handles specific errors."""
@@ -224,8 +324,7 @@ class TestErrorHandling:
                 try:
                     return handler(request)
                 except ConnectionError:
-                    fallback = AIMessage(content="Network issue, try again later")
-                    return fallback
+                    return AIMessage(content="Network issue, try again later")
 
         model = SpecificErrorModel(messages=iter([]))
         agent = create_agent(model=model, middleware=[SelectiveErrorMiddleware()])
@@ -247,8 +346,7 @@ class TestErrorHandling:
                     return result
                 except Exception:
                     call_log.append("caught-error")
-                    fallback = AIMessage(content="Recovered from error")
-                    return fallback
+                    return AIMessage(content="Recovered from error")
 
         # Test 1: Success path
         call_log.clear()
@@ -403,7 +501,6 @@ class TestStateAndRuntime:
                 for attempt in range(max_retries):
                     try:
                         return handler(request)
-                        break  # Success
                     except Exception:
                         if attempt == max_retries - 1:
                             raise
@@ -458,6 +555,49 @@ class TestMiddlewareComposition:
             "inner-before",
             "inner-after",
             "outer-after",
+        ]
+
+    def test_three_middleware_composition(self) -> None:
+        """Test composition of three middleware."""
+        execution_order = []
+
+        class FirstMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                execution_order.append("first-before")
+                response = handler(request)
+                execution_order.append("first-after")
+                return response
+
+        class SecondMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                execution_order.append("second-before")
+                response = handler(request)
+                execution_order.append("second-after")
+                return response
+
+        class ThirdMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                execution_order.append("third-before")
+                response = handler(request)
+                execution_order.append("third-after")
+                return response
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
+        agent = create_agent(
+            model=model,
+            middleware=[FirstMiddleware(), SecondMiddleware(), ThirdMiddleware()],
+        )
+
+        agent.invoke({"messages": [HumanMessage("Test")]})
+
+        # First wraps Second wraps Third: 1-before, 2-before, 3-before, model, 3-after, 2-after, 1-after
+        assert execution_order == [
+            "first-before",
+            "second-before",
+            "third-before",
+            "third-after",
+            "second-after",
+            "first-after",
         ]
 
     def test_retry_with_logging(self) -> None:
@@ -549,11 +689,9 @@ class TestMiddlewareComposition:
         class RetryMiddleware(AgentMiddleware):
             def wrap_model_call(self, request, handler):
                 try:
-                    result = handler(request)
-                    return result
+                    return handler(request)
                 except Exception:
-                    result = handler(request)
-                    return result
+                    return handler(request)
 
         class UppercaseMiddleware(AgentMiddleware):
             def wrap_model_call(self, request, handler):
@@ -570,49 +708,6 @@ class TestMiddlewareComposition:
 
         # Should retry and uppercase the result
         assert result["messages"][1].content == "SUCCESS"
-
-    def test_three_middleware_composition(self) -> None:
-        """Test composition of three middleware."""
-        execution_order = []
-
-        class FirstMiddleware(AgentMiddleware):
-            def wrap_model_call(self, request, handler):
-                execution_order.append("first-before")
-                response = handler(request)
-                execution_order.append("first-after")
-                return response
-
-        class SecondMiddleware(AgentMiddleware):
-            def wrap_model_call(self, request, handler):
-                execution_order.append("second-before")
-                response = handler(request)
-                execution_order.append("second-after")
-                return response
-
-        class ThirdMiddleware(AgentMiddleware):
-            def wrap_model_call(self, request, handler):
-                execution_order.append("third-before")
-                response = handler(request)
-                execution_order.append("third-after")
-                return response
-
-        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
-        agent = create_agent(
-            model=model,
-            middleware=[FirstMiddleware(), SecondMiddleware(), ThirdMiddleware()],
-        )
-
-        agent.invoke({"messages": [HumanMessage("Test")]})
-
-        # First wraps Second wraps Third: 1-before, 2-before, 3-before, model, 3-after, 2-after, 1-after
-        assert execution_order == [
-            "first-before",
-            "second-before",
-            "third-before",
-            "third-after",
-            "second-after",
-            "first-after",
-        ]
 
     def test_middle_retry_middleware(self) -> None:
         """Test that middle middleware doing retry causes inner to execute twice."""
@@ -674,7 +769,306 @@ class TestMiddlewareComposition:
         assert len(model_calls) == 2
 
 
-class TestAsyncOnModelCall:
+class TestWrapModelCallDecorator:
+    """Test the @wrap_model_call decorator for creating middleware."""
+
+    def test_basic_decorator_usage(self) -> None:
+        """Test basic decorator usage without parameters."""
+
+        @wrap_model_call
+        def passthrough_middleware(request, handler):
+            return handler(request)
+
+        # Should return an AgentMiddleware instance
+        assert isinstance(passthrough_middleware, AgentMiddleware)
+
+        # Should work in agent
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Hello")]))
+        agent = create_agent(model=model, middleware=[passthrough_middleware])
+
+        result = agent.invoke({"messages": [HumanMessage("Hi")]})
+        assert len(result["messages"]) == 2
+        assert result["messages"][1].content == "Hello"
+
+    def test_decorator_with_custom_name(self) -> None:
+        """Test decorator with custom middleware name."""
+
+        @wrap_model_call(name="CustomMiddleware")
+        def my_middleware(request, handler):
+            return handler(request)
+
+        assert isinstance(my_middleware, AgentMiddleware)
+        assert my_middleware.__class__.__name__ == "CustomMiddleware"
+
+    def test_decorator_retry_logic(self) -> None:
+        """Test decorator for implementing retry logic."""
+        call_count = {"value": 0}
+
+        class FailOnceThenSucceed(GenericFakeChatModel):
+            def _generate(self, messages, **kwargs):
+                call_count["value"] += 1
+                if call_count["value"] == 1:
+                    raise ValueError("First call fails")
+                return super()._generate(messages, **kwargs)
+
+        @wrap_model_call
+        def retry_once(request, handler):
+            try:
+                return handler(request)
+            except Exception:
+                # Retry once
+                return handler(request)
+
+        model = FailOnceThenSucceed(messages=iter([AIMessage(content="Success")]))
+        agent = create_agent(model=model, middleware=[retry_once])
+
+        result = agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert call_count["value"] == 2
+        assert result["messages"][1].content == "Success"
+
+    def test_decorator_response_rewriting(self) -> None:
+        """Test decorator for rewriting responses."""
+
+        @wrap_model_call
+        def uppercase_responses(request, handler):
+            result = handler(request)
+            # result is ModelResponse, extract AIMessage from it
+            ai_message = result.result[0]
+            return AIMessage(content=ai_message.content.upper())
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="hello world")]))
+        agent = create_agent(model=model, middleware=[uppercase_responses])
+
+        result = agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert result["messages"][1].content == "HELLO WORLD"
+
+    def test_decorator_error_handling(self) -> None:
+        """Test decorator for error recovery."""
+
+        class AlwaysFailModel(GenericFakeChatModel):
+            def _generate(self, messages, **kwargs):
+                raise ValueError("Model error")
+
+        @wrap_model_call
+        def error_to_fallback(request, handler):
+            try:
+                return handler(request)
+            except Exception:
+                return AIMessage(content="Fallback response")
+
+        model = AlwaysFailModel(messages=iter([]))
+        agent = create_agent(model=model, middleware=[error_to_fallback])
+
+        result = agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert result["messages"][1].content == "Fallback response"
+
+    def test_decorator_with_state_access(self) -> None:
+        """Test decorator accessing agent state."""
+        state_values = []
+
+        @wrap_model_call
+        def log_state(request, handler):
+            state_values.append(request.state.get("messages"))
+            return handler(request)
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
+        agent = create_agent(model=model, middleware=[log_state])
+
+        agent.invoke({"messages": [HumanMessage("Test")]})
+
+        # State should contain the user message
+        assert len(state_values) == 1
+        assert len(state_values[0]) == 1
+        assert state_values[0][0].content == "Test"
+
+    def test_multiple_decorated_middleware(self) -> None:
+        """Test composition of multiple decorated middleware."""
+        execution_order = []
+
+        @wrap_model_call
+        def outer_middleware(request, handler):
+            execution_order.append("outer-before")
+            result = handler(request)
+            execution_order.append("outer-after")
+            return result
+
+        @wrap_model_call
+        def inner_middleware(request, handler):
+            execution_order.append("inner-before")
+            result = handler(request)
+            execution_order.append("inner-after")
+            return result
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
+        agent = create_agent(model=model, middleware=[outer_middleware, inner_middleware])
+
+        agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert execution_order == [
+            "outer-before",
+            "inner-before",
+            "inner-after",
+            "outer-after",
+        ]
+
+    def test_decorator_with_custom_state_schema(self) -> None:
+        """Test decorator with custom state schema."""
+        from typing_extensions import TypedDict
+
+        class CustomState(TypedDict):
+            messages: list
+            custom_field: str
+
+        @wrap_model_call(state_schema=CustomState)
+        def middleware_with_schema(request, handler):
+            return handler(request)
+
+        assert isinstance(middleware_with_schema, AgentMiddleware)
+        # Custom state schema should be set
+        assert middleware_with_schema.state_schema == CustomState
+
+    def test_decorator_with_tools_parameter(self) -> None:
+        """Test decorator with tools parameter."""
+        from langchain_core.tools import tool
+
+        @tool
+        def test_tool(query: str) -> str:
+            """A test tool."""
+            return f"Result: {query}"
+
+        @wrap_model_call(tools=[test_tool])
+        def middleware_with_tools(request, handler):
+            return handler(request)
+
+        assert isinstance(middleware_with_tools, AgentMiddleware)
+        assert len(middleware_with_tools.tools) == 1
+        assert middleware_with_tools.tools[0].name == "test_tool"
+
+    def test_decorator_parentheses_optional(self) -> None:
+        """Test that decorator works both with and without parentheses."""
+
+        # Without parentheses
+        @wrap_model_call
+        def middleware_no_parens(request, handler):
+            return handler(request)
+
+        # With parentheses
+        @wrap_model_call()
+        def middleware_with_parens(request, handler):
+            return handler(request)
+
+        assert isinstance(middleware_no_parens, AgentMiddleware)
+        assert isinstance(middleware_with_parens, AgentMiddleware)
+
+    def test_decorator_preserves_function_name(self) -> None:
+        """Test that decorator uses function name for class name."""
+
+        @wrap_model_call
+        def my_custom_middleware(request, handler):
+            return handler(request)
+
+        assert my_custom_middleware.__class__.__name__ == "my_custom_middleware"
+
+    def test_decorator_mixed_with_class_middleware(self) -> None:
+        """Test decorated middleware mixed with class-based middleware."""
+        execution_order = []
+
+        @wrap_model_call
+        def decorated_middleware(request, handler):
+            execution_order.append("decorated-before")
+            result = handler(request)
+            execution_order.append("decorated-after")
+            return result
+
+        class ClassMiddleware(AgentMiddleware):
+            def wrap_model_call(self, request, handler):
+                execution_order.append("class-before")
+                result = handler(request)
+                execution_order.append("class-after")
+                return result
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
+        agent = create_agent(
+            model=model,
+            middleware=[decorated_middleware, ClassMiddleware()],
+        )
+
+        agent.invoke({"messages": [HumanMessage("Test")]})
+
+        # Decorated is outer, class-based is inner
+        assert execution_order == [
+            "decorated-before",
+            "class-before",
+            "class-after",
+            "decorated-after",
+        ]
+
+    def test_decorator_complex_retry_logic(self) -> None:
+        """Test decorator with complex retry logic and backoff."""
+        attempts = []
+        call_count = {"value": 0}
+
+        class UnreliableModel(GenericFakeChatModel):
+            def _generate(self, messages, **kwargs):
+                call_count["value"] += 1
+                if call_count["value"] <= 2:
+                    raise ValueError(f"Attempt {call_count['value']} failed")
+                return super()._generate(messages, **kwargs)
+
+        @wrap_model_call
+        def retry_with_tracking(request, handler):
+            max_retries = 3
+            for attempt in range(max_retries):
+                attempts.append(attempt + 1)
+                try:
+                    return handler(request)
+                except Exception:
+                    # On error, continue to next attempt
+                    if attempt < max_retries - 1:
+                        continue  # Retry
+                    else:
+                        raise  # All retries failed
+
+        model = UnreliableModel(messages=iter([AIMessage(content="Finally worked")]))
+        agent = create_agent(model=model, middleware=[retry_with_tracking])
+
+        result = agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert attempts == [1, 2, 3]
+        assert result["messages"][1].content == "Finally worked"
+
+    def test_decorator_request_modification(self) -> None:
+        """Test decorator modifying request before execution."""
+        modified_prompts = []
+
+        @wrap_model_call
+        def add_system_prompt(request, handler):
+            # Modify request to add system prompt
+            modified_request = ModelRequest(
+                messages=request.messages,
+                model=request.model,
+                system_prompt="You are a helpful assistant",
+                tool_choice=request.tool_choice,
+                tools=request.tools,
+                response_format=request.response_format,
+                state={},
+                runtime=None,
+            )
+            modified_prompts.append(modified_request.system_prompt)
+            return handler(modified_request)
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Response")]))
+        agent = create_agent(model=model, middleware=[add_system_prompt])
+
+        agent.invoke({"messages": [HumanMessage("Test")]})
+
+        assert modified_prompts == ["You are a helpful assistant"]
+
+
+class TestAsyncWrapModelCall:
     """Test async execution with wrap_model_call."""
 
     async def test_async_model_with_middleware(self) -> None:
@@ -686,7 +1080,6 @@ class TestAsyncOnModelCall:
                 log.append("before")
                 result = await handler(request)
                 log.append("after")
-
                 return result
 
         model = GenericFakeChatModel(messages=iter([AIMessage(content="Async response")]))
@@ -723,6 +1116,92 @@ class TestAsyncOnModelCall:
         assert call_count["value"] == 2
         assert result["messages"][1].content == "Async success"
 
+    async def test_decorator_with_async_agent(self) -> None:
+        """Test that decorated middleware works with async agent invocation."""
+        call_log = []
+
+        @wrap_model_call
+        async def logging_middleware(request, handler):
+            call_log.append("before")
+            result = await handler(request)
+            call_log.append("after")
+            return result
+
+        model = GenericFakeChatModel(messages=iter([AIMessage(content="Async response")]))
+        agent = create_agent(model=model, middleware=[logging_middleware])
+
+        result = await agent.ainvoke({"messages": [HumanMessage("Test")]})
+
+        assert call_log == ["before", "after"]
+        assert result["messages"][1].content == "Async response"
+
+
+class TestSyncAsyncInterop:
+    """Test sync/async interoperability."""
+
+    def test_sync_invoke_with_only_async_middleware_raises_error(self) -> None:
+        """Test that sync invoke with only async middleware raises error."""
+
+        class AsyncOnlyMiddleware(AgentMiddleware):
+            async def awrap_model_call(
+                self,
+                request: ModelRequest,
+                handler: Callable[[ModelRequest], Awaitable[AIMessage]],
+            ) -> AIMessage:
+                return await handler(request)
+
+        agent = create_agent(
+            model=FakeToolCallingModel(),
+            tools=[],
+            system_prompt="You are a helpful assistant.",
+            middleware=[AsyncOnlyMiddleware()],
+        )
+
+        with pytest.raises(NotImplementedError):
+            agent.invoke({"messages": [HumanMessage("hello")]})
+
+    def test_sync_invoke_with_mixed_middleware(self) -> None:
+        """Test that sync invoke works with mixed sync/async middleware when sync versions exist."""
+        calls = []
+
+        class MixedMiddleware(AgentMiddleware):
+            def before_model(self, state, runtime) -> None:
+                calls.append("MixedMiddleware.before_model")
+
+            async def abefore_model(self, state, runtime) -> None:
+                calls.append("MixedMiddleware.abefore_model")
+
+            def wrap_model_call(
+                self,
+                request: ModelRequest,
+                handler: Callable[[ModelRequest], AIMessage],
+            ) -> AIMessage:
+                calls.append("MixedMiddleware.wrap_model_call")
+                return handler(request)
+
+            async def awrap_model_call(
+                self,
+                request: ModelRequest,
+                handler: Callable[[ModelRequest], Awaitable[AIMessage]],
+            ) -> AIMessage:
+                calls.append("MixedMiddleware.awrap_model_call")
+                return await handler(request)
+
+        agent = create_agent(
+            model=FakeToolCallingModel(),
+            tools=[],
+            system_prompt="You are a helpful assistant.",
+            middleware=[MixedMiddleware()],
+        )
+
+        result = agent.invoke({"messages": [HumanMessage("hello")]})
+
+        # In sync mode, only sync methods should be called
+        assert calls == [
+            "MixedMiddleware.before_model",
+            "MixedMiddleware.wrap_model_call",
+        ]
+
 
 class TestEdgeCases:
     """Test edge cases and error conditions."""
@@ -753,12 +1232,10 @@ class TestEdgeCases:
             def wrap_model_call(self, request, handler):
                 attempts.append("first-attempt")
                 try:
-                    result = handler(request)
-                    return result
+                    return handler(request)
                 except Exception:
                     attempts.append("retry-attempt")
-                    result = handler(request)
-                    return result
+                    return handler(request)
 
         call_count = {"value": 0}
