+++ libs/langchain_v1/tests/unit_tests/agents/middleware/implementations/test_shell_tool.py
